Enable the switch of __riscv_hwprobe() from unsigned long* to cpu_set_t*.
glibc wants to go with a transparent union (which only works in C),
but I think we should just fix the one existing user (this project)
and move on with our lives. In the meantime, this lets us fix bionic
and prevent new instances while we work on the upstream fix.

Change-Id: Ice06728fda5dece3f0695c3469765e9b85c39a64